Reorg tests and renamed run.py
